Move Search to its own component
